artifacts/manhwa-tracker/requirements.txt
discord.py>=2.3.2
cloudscraper>=1.2.71
beautifulsoup4>=4.12.3
lxml>=5.2.2
requests>=2.31.0
Pillow>=10.3.0
google-auth>=2.29.0
google-api-python-client>=2.125.0
aiohttp>=3.9.5
